Sharing test preparation
