+++ src/test/java/buttondevteam/core/PlayerDataTest.java
@@ -1,15 +1,6 @@
 package buttondevteam.core;
 
 import java.util.UUID;
-import java.util.logging.Logger;
-
-import org.bukkit.Bukkit;
-import org.bukkit.Server;
-import org.bukkit.plugin.PluginManager;
-import org.mockito.Mockito;
-import org.mockito.invocation.InvocationOnMock;
-import org.mockito.stubbing.Answer;
-
 import buttondevteam.core.TestPlayerClass.TestEnum;
 import buttondevteam.lib.player.TBMCPlayerBase;
 import junit.framework.Test;
@@ -29,21 +20,7 @@ public class PlayerDataTest extends TestCase {
 	}
 
 	public void testConfig() {
-		Bukkit.setServer(Mockito.mock(Server.class, new Answer<Object>() {
-
-			@Override
-			public Object answer(InvocationOnMock invocation) throws Throwable {
-				// System.out.println("Return type: " + invocation.getMethod().getReturnType());
-				// System.out.println(String.class.isAssignableFrom(invocation.getMethod().getReturnType()));
-				if (String.class.isAssignableFrom(invocation.getMethod().getReturnType()))
-					return "";
-				if (Logger.class.isAssignableFrom(invocation.getMethod().getReturnType()))
-					return Logger.getAnonymousLogger();
-				if (PluginManager.class.isAssignableFrom(invocation.getMethod().getReturnType()))
-					return Mockito.mock(PluginManager.class);
-				return null;
-			}
-		}));
+		TestPrepare.PrepareServer();
 		UUID uuid = new UUID(0L, 0L);
 		try (TestPlayerClass p = TBMCPlayerBase.getPlayer(uuid, TestPlayerClass.class)) {
 			p.PlayerName().set("Test");
